test(restaking): refactor testing layout and added basic reward harvesting test
