+++ programs/restaking/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "restaking"
-version = "0.6.0"
+version = "0.6.1"
 description = "Fragmetric Liquid Restaking Program"
 authors = { workspace = true }
 repository = { workspace = true }
