+++ programs/testutil/cli.executable.ts
@@ -1,16 +1,19 @@
 #!/usr/bin/env pnpm tsx --no-warnings
 
-import { initializeFragBTC } from '../restaking/tests/fragbtc';
-import { initializeFragJTO } from '../restaking/tests/fragjto';
-import { initializeFragSOL } from '../restaking/tests/fragsol';
+import { initializeFragBTC } from '../restaking/tests/fragbtc.init';
+import { initializeFragJTO } from '../restaking/tests/fragjto.init';
+import { initializeFragSOL } from '../restaking/tests/fragsol.init';
 import { createTestSuiteContext } from './context';
 
 createTestSuiteContext({ validator: 'litesvm' })
   .then(async (ctx) => {
     // init receipt tokens sequentially
-    ctx.sdk.logger.start('Initialize fragSOL...');
-    return initializeFragSOL(ctx)
-      .initializationTasks.then(() => {
+    return Promise.resolve()
+      .then(() => {
+        ctx.sdk.logger.start('Initialize fragSOL...');
+        return initializeFragSOL(ctx).initializationTasks;
+      })
+      .then(() => {
         ctx.sdk.logger.start('Initialize fragJTO...');
         return initializeFragJTO(ctx).initializationTasks;
       })
