+++ programs/testutil/vitest.ts
@@ -34,6 +34,7 @@ export async function masked(
     /[.*S|s]lots?$/,
     /.*At?$/,
     'signature',
+    'contribution',
     ...(options?.extraKeys ?? []),
   ];
   const values = options?.values ?? [
